doc: Rework everything
Basically rewrite all the documentation. The usage guide is
significantly improved, adding a detailed comparison of drivers,
features and implementations and examples for how this can actually be
used as a directive (something that wasn't at all obvious at first
glance).

We also take the opportunity to change the new-style documentation
layout.

Change-Id: I7fcd49f580450e74f716fc5b388e5370deca4b04
Signed-off-by: Stephen Finucane <stephen@that.guru>