Handle ConfigParser.readfp deprecation
The readfp call was deprecated in Python 3.2 in favor of the new
read_file call. This call did not exist in Python 2, so in order to get
rid of deprecation warnings when building feature matrix output, this
aliases read_file in Python 2 to readfp but uses the new call when
supported.

Change-Id: Ia67752ddd898374983c0de38b97122aa31e98441
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>