Report sphinx extension is thread safe
This updates the extension to return values to indicate it can be run
multithreaded. This allows consumers of the extension to use multiple
threads when building documentation, since sphinx either requires all
projects are thread safe, or else it emits a warning that many of our
projects then fail the build on with treating warnings as errors.

Change-Id: Ie7ce3a3d1c592debf302117fce2c0f3fae0b19a2
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>