Cleanup misplaced release note
A release note was added in the wrong location, so it was missed for
that release. Since moving it to the correct location now would cause it
to show up in the next release's release notes, potentially causing
confusion, this just removes the note.

Added over a year ago in:
I0f2bdee6eb67753f5f365b2681c4c7e3ae35c3ec

Change-Id: If9e97bfea07b66a213e39fc9b4877fde9f8f30e5
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>